+++ docs/sandbox/Examples.md
@@ -11,6 +11,7 @@
 
 - Initial setup of the first new OTP Sandbox Extension
 - Added a simple GraphQL API for retrieving Graph statistics
+- Moved Graph Example Updaters from main code to sandbox examples. 
 
 
 ## Documentation
@@ -20,3 +21,11 @@ This extension show how to create a web endpoint to get some simple statistics:
 - Number of stops in the graph
 - Number of stations in the graph
  
+### Graph Example Updaters
+There is two example updaters, see the Java doc for more information on how to implement 
+new updater. 
+- The `ExampleGraphUpdater` class shows an example of how to implement a graph updater suited for 
+  streaming updaters.
+- The `ExamplePollingGraphUpdater` class shows an example of how to implement a graph updater 
+   suited for polling updaters.
+
